% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function.R
\name{cost_function}
\alias{cost_function}
\title{Cost function}
\usage{
cost_function(x, y, theta)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}

\item{theta}{a matrix}
}
\value{
a matrix
}
\description{
Cost function
}
\examples{
x <- as.matrix(breast_cancer[,1:4])
y <- as.matrix(ifelse(breast_cancer$classe =="malignant",1,0))
theta <- as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1))
cost_function(x,y,theta)
}
