% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_plot.R
\name{cost_plot}
\alias{cost_plot}
\title{Cost_plot function}
\usage{
cost_plot(Reg.Log)
}
\arguments{
\item{Reg.Log}{object of the S3 class Reg.log provide by the rlgd.fit function}
}
\value{
a graphic of the cost for each iteration of the selected method of gradient
}
\description{
Display a graphic of the cost for each iteration of the selected method of gradient.
}
\examples{
library(caret)
train_index <- createDataPartition(breast_cancer$classe, p = 0.7, list = FALSE)
train_set <- breast_cancer[train_index, ]
res <- rlgd.fit(classe ~ ., train_set, mode = "batch", learning_rate = 0.01,
max_iter = 200, tol = 1e-4, ncores=1)
cost_plot(res)
}
