% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient of the cost function of logistic regression}
\usage{
gradient(x, y, theta)
}
\arguments{
\item{x}{matrix of features of dimension m x n}

\item{y}{matrix of target of dimension m x 1}

\item{theta}{matrix of parameters of dimension n x 1}
}
\value{
matrix of dimension n x 1 of partial derivatives of each parameter
}
\description{
Gradient of the cost function of logistic regression
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
y <- breast_cancer[1:10,"classe"]
y <- as.matrix(ifelse(y$classe =="malignant",1,0))
theta <- as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1))
gradient(x,y,theta)
}
