% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_parallele.R
\name{gradient_parallele}
\alias{gradient_parallele}
\title{Gradient function for the parallel use}
\usage{
gradient_parallele(X, theta)
}
\arguments{
\item{theta}{}
}
\value{

}
\description{
Gradient function for the parallel use
}
\examples{
gradient_parallele(blocs[[1]], as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1)))
}
