% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blocs.R
\name{make_blocs}
\alias{make_blocs}
\title{Creation of blocs for parallelisation}
\usage{
make_blocs(x, y, ncores)
}
\arguments{
\item{x}{matrix of features variable}

\item{y}{matrix of target variable}

\item{ncores}{integer that indicate how much hearts to use pr the parallelisation}
}
\value{
blocs, a list of several blocs (one for each used heart) containing parts of x and y
}
\description{
The goal here is to create an input for the parSapply in the parallelisation function. It requires several blocs to used the parallelisaton on. In our case, each blocs is composed of a part of x and y, the total give the complete x and y
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
y <- breast_cancer[1:10,"classe"]
y <- as.matrix(ifelse(y$classe =="malignant",1,0))
make_blocs(x,y,ncores=3)
}
