% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelisation.R
\name{parallelisation}
\alias{parallelisation}
\title{parallelisation function}
\usage{
parallelisation(blocs, fonction, theta, clust)
}
\arguments{
\item{blocs}{set of data}

\item{fonction}{function to parallelize}

\item{theta}{matrix}

\item{clust}{cluster of blocks}
}
\value{
sum of gradients of each blocs
}
\description{
parallelisation function
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
y <- breast_cancer[1:10,"classe"]
y <- as.matrix(ifelse(y$classe =="malignant",1,0))
blocs <- make_blocs(x,y,ncores = 3)
}
