% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Reg.Log}
\alias{print.Reg.Log}
\title{Overload of print function for x of class Reg.log}
\usage{
\method{print}{Reg.Log}(x, ...)
}
\arguments{
\item{x}{of S3 class Reg.log}

\item{...}{other parameters}
}
\value{
to do
}
\description{
Overload of print function for x of class Reg.log
}
\examples{
res <- rlgd.fit(classe ~ .,breast_cancer,"batch",batch_size = 32,
learning_rate = 0.5, max_iter = 100, tol = 1e-4, ncores = 1, standardize = FALSE)
print(res)
}
