% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Probability function}
\usage{
probability(x, theta)
}
\arguments{
\item{x}{matrix of dim (m,n+1) where n is the number of features}

\item{theta}{matrix of dim (n+1,1) of parameters}
}
\value{
1 / (1 + exp(-x\%*\%theta))
}
\description{
Function used to compute the probability that observations are in a specified category of a binary variable.
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
theta <- as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1))
probability(x,theta)
}
\seealso{
\code{\link{sigmoid}} \code{\link{x_dot_theta}}
}
