% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{rlgd.fit}
\alias{rlgd.fit}
\title{Fit Binary Logistic Regression using gradient descent}
\usage{
rlgd.fit(
  formula,
  data,
  mode,
  batch_size = 16,
  learning_rate = 0.1,
  max_iter = 100,
  tol = 1e-04,
  ncores = 1,
  standardize = FALSE
)
}
\arguments{
\item{formula}{object of class formula e.g y ~ a + b}

\item{data}{a data frame}

\item{mode}{the gradient descent method used to fit the model : "batch", "online" or "mini-batch"}

\item{batch_size}{a integer used to fix the batch size of mini-batch gradient descent.}

\item{learning_rate}{a float. Learning rate of the gradient descent.}

\item{max_iter}{an integer. Number of maximum number of iteration/epoch of the gradient descent}

\item{tol}{a numeric. Set to 1e-4 by default. Threshold to stop iteration.}

\item{ncores}{only used for batch mode. an integer corresponding to the number of cores used during parallelisation process}

\item{standardize}{logical. FALSE by default, set to TRUE to apply standardization.}
}
\value{
an instance of the class Reg.log
}
\description{
This function perform binary logistic regression using gradient descent optimization (batch, online or mini-batch). It can use
parallel computing for batch mode.
}
\details{
Test ddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd
dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd
ddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd

TES32 ddddddddddddddddddddddddddddddddddddddddddddddddddddd
dddddddddddddddddddddddddddddddddddddddddddddddddddddddddd
ddddddddddddddddddddddddddddddddddddddddddddddddddddd
}
\examples{
rlgd.fit(classe ~ .,breast_cancer,"batch",batch_size = 32,
learning_rate = 0.5, max_iter = 100, tol = 1e-4, ncores = 2, standardize = TRUE)
}
