% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{rlgd.predict}
\alias{rlgd.predict}
\title{Predict function}
\usage{
rlgd.predict(Reg.log, newdata, type)
}
\arguments{
\item{Reg.log}{variable containing the result of the fit function}

\item{newdata}{matrix}

\item{type}{string ('class' or 'posterior')}
}
\value{
predicted classes or probabilites to belong to classes
}
\description{
Predict function
}
\examples{
set.seed(10)
library(caret)
train_index <- createDataPartition(breast_cancer$classe, p = 0.7, list = FALSE)
train_set <- breast_cancer[train_index, ]
test_set <- breast_cancer[-train_index, ]
res <- rlgd.fit(classe ~ ., train_set, mode = "batch", batch_size = 32, learning_rate = 0.01,
max_iter = 100, tol = 1e-4, standardize = FALSE)
rlgd.predict(res, test_set[, res$x_names], type = "class")

}
