% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Reg.Log}
\alias{summary.Reg.Log}
\title{Overload of summary function for object of class Reg.Log}
\usage{
\method{summary}{Reg.Log}(object, ...)
}
\arguments{
\item{object}{of S3 class Reg.log}

\item{...}{other parameters}
}
\value{
to do
}
\description{
Overload of summary function for object of class Reg.Log
}
\examples{
res <- rlgd.fit(classe ~ .,breast_cancer,"batch",batch_size = 32,
learning_rate = 0.5, max_iter = 100, tol = 1e-4, ncores = 1, standardize = FALSE)
summary(res)
}
