% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_gradient_descent.R
\name{batch_gradient_descent}
\alias{batch_gradient_descent}
\title{Gradient Descent Function - Batch}
\usage{
batch_gradient_descent(X, theta, learning_rate, n_iter)
}
\arguments{
\item{X}{a matrix}

\item{theta}{a matrix}

\item{learning_rate}{a float}

\item{n_iter}{a integer}
}
\value{
list of matrix
}
\description{
Gradient Descent Function - Batch
}
\examples{
batch_gradient_descent(blocs,as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1)),0.01,100)
}
