% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_gradient_descent.R
\name{batch_gradient_descent}
\alias{batch_gradient_descent}
\title{Gradient Descent Function - Batch}
\usage{
batch_gradient_descent(x, y, theta, learning_rate, max_iter, tol, ncores)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}

\item{theta}{a matrix}

\item{learning_rate}{a float}

\item{max_iter}{a integer}

\item{tol}{a numeric}

\item{ncores}{a integer}
}
\value{
a list containing 2 matrices : matrix of theta and matrix of cost history
}
\description{
Gradient Descent Function - Batch
}
\examples{
batch_gradient_descent(x,y,as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1)),0.01,100,1e-4,ncores=3)
}
