% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_x_y.R
\name{get_x_y}
\alias{get_x_y}
\title{get_x_y_function}
\usage{
get_x_y(formula, data, standardize)
}
\arguments{
\item{formula}{R Formula e.g y ~ x + b}

\item{data}{dataframe including features and target variable}

\item{standardize}{TRUE or FALSE}
}
\value{
list(y,x,y_name,x_name), y & x = matrix, y_name & x_names = c()
}
\description{
get_x_y_function
}
\examples{
get_x_y(classe ~.,breast_cancer, standardize=TRUE)
}
