% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Gradient function}
\usage{
gradient(x, y, theta)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}

\item{theta}{a matrix}
}
\value{
a matrix
}
\description{
Gradient function
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
y <- breast_cancer[1:10,"classe"]
y <- as.matrix(ifelse(y$classe =="malignant",1,0))
theta <- as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1))
gradient(x,y,theta)
}
