% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_parallele.R
\name{gradient_parallele}
\alias{gradient_parallele}
\title{gradient parallele}
\usage{
gradient_parallele(X, theta)
}
\arguments{
\item{X}{matrix}

\item{theta}{matrix}
}
\value{
gradient
}
\description{
gradient parallele
}
