% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_parallele.R
\name{gradient_parallele}
\alias{gradient_parallele}
\title{Gradient parallele}
\usage{
gradient_parallele(X, theta)
}
\arguments{
\item{X}{list of several blocks (one for each used heart) containing parts of x and y}

\item{theta}{matrix of parameters}
}
\value{
the calculated gradient
}
\description{
This function is a simple copy of the gradient function but with the X parameters that replace x and y.
}
\details{
It was done like that because the parSapply used in the parallelisation function require in the input the list of the blocs used to parallelize the function. Because of that, we had to merge x and y in one block.
}
