% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelisation.R
\name{parallelisation}
\alias{parallelisation}
\title{Parallelisation function}
\usage{
parallelisation(blocs, fonction, theta, clust)
}
\arguments{
\item{blocs}{list of several blocs (one for each used heart) containing parts of x and y}

\item{fonction}{function to parallelize, in our case the gradient_parallele function}

\item{theta}{matrix of parameters}

\item{clust}{cluster of nodes for parallel computation (one for each used heart)}
}
\value{
sum of gradients applied to each blocs
}
\description{
The aim of parallelization is to speed up the calculation time of an algorithm by making the most of the increased capacity of multicore processors.
}
\details{
The role of the parSapply function is to execute the selected function on each blocs at the same time (parallelisation).
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
y <- breast_cancer[1:10,"classe"]
y <- as.matrix(ifelse(y$classe =="malignant",1,0))
blocs <- make_blocs(x,y,ncores = 3)
}
