% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelisation.R
\name{parallelisation}
\alias{parallelisation}
\title{Parallelisation function}
\usage{
parallelisation(x, y, ncores, fonction, theta)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}

\item{ncores}{a numeric}

\item{fonction}{a function}

\item{theta}{a matrix}
}
\value{
the result return by the function used in the parallelisation
}
\description{
Parallelisation function
}
\examples{
parallelisation(x,y,ncores = 3,gradient_parallele, as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1)))
}
