% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Probability function of logistic regression}
\usage{
probability(x, theta)
}
\arguments{
\item{x}{matrix of dimension m x n}

\item{theta}{matrix of dimension n x 1 of parameters}
}
\value{
matrix of dimension m x 1 equal to (1 / (1 + exp(-x\%*\%theta)))
}
\description{
Function used to compute the probability that observations are in a specified category of a binary variable.
}
\examples{
x <- as.matrix(breast_cancer[1:10,1:4])
theta <- as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1))
probability(x,theta)
}
\seealso{
\itemize{
\item \code{\link{sigmoid}}
\item \code{\link{x_dot_theta}}
}
}
