% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{a number}
}
\value{
1 / (1 + exp(-x))
}
\description{
This function is used to perform logistic regression
}
\examples{
sigmoid(2)
}
