% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_dot_theta.R
\name{x_dot_theta}
\alias{x_dot_theta}
\title{x_dot_theta function}
\usage{
x_dot_theta(x, theta)
}
\arguments{
\item{x}{a matrix}

\item{theta}{a matrix}
}
\value{
a matrix
}
\description{
x_dot_theta function
}
\examples{
x_dot_theta(x,as.matrix(rnorm(n = dim(x)[2], mean = 0, sd = 1)))
}
